requirements/base.txt
boto3
Django>=3.2.9,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.9,<2.10.0
Markdown>=3.3.5,<3.4.0
django-filter>=21.1,<21.2.0
flake8>=4.0.1,<4.1.0
pyvips>=2.1.16,<2.2.0
python-magic>=0.4.24,<0.5.0
